Fix styling and file structure
