+++ Views/Dashboard.js
@@ -1,25 +1,9 @@
-import { StatusBar } from 'expo-status-bar';
 import React from 'react';
-import { StyleSheet, Text, View, TouchableOpacity } from 'react-native';
+import { StyleSheet, Text, View } from 'react-native';
 //Libraries
-import * as ImagePicker from 'expo-image-picker';
-import { Button } from 'react-native-paper';
 import PickImage from "../components/PickImage";
 
-
 export default function Dashboard() {
-
-  // Image picker init
-  let openImagePickerAsync = async () => {
-    console.log("u been clicked n picked bro")
-    let permissionResult = await ImagePicker.requestMediaLibraryPermissionsAsync();
-    if (!permissionResult.granted) {
-      alert("Permission to access camera roll is required!");
-      return;
-    }
-    let pickerResult = await ImagePicker.launchImageLibraryAsync();
-    console.log("u got pickered", pickerResult);
-  }
   return (
       <View>
         <Text>Dashboard</Text>
