+++ components/PickImageStyles.js
@@ -1,0 +1,27 @@
+import { StyleSheet } from 'react-native'
+
+const styles = StyleSheet.create({
+    container: {
+      flex: 1,
+      backgroundColor: '#fff',
+      alignItems: 'center',
+      justifyContent: 'center',
+    },
+    button: {
+      backgroundColor: "#e4e4e4",
+      alignItems: "center",
+      justifyContent: "center",
+      borderColor: "#BFBFBF",
+      borderWidth: 1,
+      elevation: 5,
+      marginTop: 20,
+      width: 210,
+      height: 70,
+    },
+    thumbnail: {
+      height: 200,
+      width: 200,
+    }
+  });
+
+  export default styles
